rename serializers with capital S to lower s
